==> experiments/exp-mmrf/00_02_clean_gen_exp.ipynb <==
# Loading Data Set

import pandas as pd

# Formatting Gene Expressions

# loading gene counts
gene_fpkm = pd.read_csv('data/gene_count.txt', sep='\t', index_col='GENE_ID')

# removing data not collected at the first trail
for col in gene_fpkm.columns:
    if '_1_' not in col:
        del gene_fpkm[col]

# transpose matrix, delete patients and gene with all nan, and replace remainder missing by zero
gene_fpkm = gene_fpkm.T.dropna(how='all', axis=0).dropna(how='any', axis=1)

# replace id column name
gene_fpkm.index.name = 'ID'

# normalize index value transforming mmrf ids to integers
gene_fpkm.index = [''.join(col.split('_')[:2]) for col in gene_fpkm.index]

# selected class
# gene_details = pd.read_csv('data/gene_details.tsv', sep='\t')
# gene_selected_class = pd.read_csv('data/gene_selected_class.tsv', sep='\t')
# gene_selected_class = gene_details.merge(gene_selected_class, on='gene_biotype').set_index('ensembl_gene_id')
# gene_selected_class = [gen for gen in gene_selected_class.index if gen in gene_fpkm.columns]
# gene_fpkm = gene_fpkm[gene_selected_class]

# removing genes with zero sum
gene_fpkm = gene_fpkm[list(gene_fpkm.sum()[gene_fpkm.sum() > 0].index)]

# removing duplicated index
gene_fpkm = gene_fpkm.loc[~gene_fpkm.index.duplicated(keep='first')]

# removing duplicated columns
gene_fpkm = gene_fpkm.loc[:,~gene_fpkm.columns.duplicated()]

gene_fpkm.index.name = 'ID'

gene_fpkm.to_csv('data/gene_count.tsv', sep='\t')

print(gene_fpkm.shape)

gene_fpkm.iloc[:8,:8]
