==> experiments/exp-mmrf/00_03_clean_per_response.ipynb <==
from collections import Counter

import pandas as pd
import numpy as np

clinical = pd.read_csv('data/clinical.tsv', sep='\t', index_col='ID')

genefpkm = pd.read_csv('data/gene_count.tsv', sep='\t', index_col='ID')

pd.DataFrame({'Rows': [genefpkm.shape[0], clinical.shape[0]], 
              'Columns': [genefpkm.shape[1], clinical.shape[1]]}, index=['GENEFPKM', 'CLINICAL'])

joined = clinical.join(genefpkm, how='inner')

pd.DataFrame({'Rows': [joined.shape[0]], 
              'Columns': [joined.shape[1]]}, index=['JOINED'])

pd.DataFrame(dict(Counter(joined['therapy_first_line_class'])), index=['Count'])

pd.DataFrame(Counter(joined['response_best_response_first_line'].replace(np.nan, 'nan')), index=['Count'])

pd.DataFrame(Counter(joined['response_days_to_disease_progression'].replace(np.nan, 'nan')), index=['Count'])

df_dtdp = clinical.copy()

del df_dtdp['response_best_response_first_line']

df_dtdp = df_dtdp.loc[~df_dtdp['response_days_to_disease_progression'].isnull()]

df_dtdp.to_csv('data/clinical_dtdp.tsv', sep='\t', index=True)

pd.DataFrame(Counter(df_dtdp['response_days_to_disease_progression'].replace(np.nan, 'nan')), index=['Count'])

df_brfl = clinical.copy()

del df_brfl['response_days_to_disease_progression']

df_brfl = df_brfl.loc[~df_brfl['response_best_response_first_line'].isnull()]

df_brfl.to_csv('data/clinical_brfl.tsv', sep='\t', index=True)

pd.DataFrame(Counter(df_brfl['response_best_response_first_line'].replace(np.nan, 'nan')), index=['Count'])

pd.DataFrame(Counter(df_brfl['therapy_first_line_class']), index=['Count'])

pd.DataFrame(Counter(df_dtdp['therapy_first_line_class']), index=['Count'])
