==> experiments/exp-mmrf/02_02_gene_contrib_gene.ipynb <==
# Loading Clinical Data

from data import load_data

clinical, _, genes, treatments, outcome = load_data()

clinical.head()

from evaluation import optimize_threshold, classification_metrics
from constants import N_FOLDS, RANDOM_STATE

from correlation import select_genes

from sklearn.metrics import roc_auc_score, log_loss, confusion_matrix
from sklearn.model_selection import StratifiedKFold

from optimization import LightGBMOptimizer

from lightgbm import LGBMModel

import pandas as pd
import pickle
import time
import os

result = {c: [] for c in ['experiment', 'marker', 'train_auc', 'valid_auc', 
                          'train_loss', 'valid_loss', 'execution_time', 'threshold']}

selected_genes_mem = {}

for t in clinical.columns:
    
    print('*********************************************************************************')
    print(t)
    print('*********************************************************************************\n')
    
    y = outcome.dropna()
    
    x = clinical[[t]].dropna().join(treatments, how='inner')
    
    x = x.join(genes).join(y)[x.columns]
    y = x.join(y)[y.columns[0]]
    
    opt_kf = StratifiedKFold(N_FOLDS, shuffle=True, random_state=RANDOM_STATE)
    
    if y.sum() < 10 or (1 - y).sum() < 10:
        print('Ignoring {} marker once it has less than five treatment senstive patients associted.')
    
    for experiment, (opt_train, opt_valid) in enumerate(opt_kf.split(x, y)):
        
        initial_time = time.time()
        
        #################################################################################################
        # Train and Valid Split
        #################################################################################################
        
        genes_train = genes.iloc[opt_train, :]
        genes_valid = genes.iloc[opt_valid, :]
        
        x_train, y_train = x.iloc[opt_train, :], y.iloc[opt_train]
        x_valid, y_valid = x.iloc[opt_valid, :], y.iloc[opt_valid]
        
        #################################################################################################
        # Gene Expression Selection
        #################################################################################################        
        
        if experiment not in selected_genes_mem:
            selected_genes_mem[experiment] = select_genes(genes_train, y_train, threshold=0.002)
            
        selected_genes = selected_genes_mem[experiment][0]
        
        genes_train = genes_train[selected_genes]
        genes_valid = genes_valid[selected_genes]
                                  
        x_train = x_train.join(genes_train)
        x_valid = x_valid.join(genes_valid)
        
        #################################################################################################
        # Hyper parameters optimization
        #################################################################################################        
        
        training_default_params = {
            'objective': 'binary',
            'metric': 'binary_logloss',
            'n_estimators': 100,
            'objective': 'binary',
            'is_unbalance': False, 
            'extra_trees': True,
            'max_depth': 4,
            'learning_rate': 0.1,
            'min_split_gain': 0.0001,
            'min_child_weight': 0.0001,
            'random_state': RANDOM_STATE}
        
        optimizer_params = {
            'n_folds': 2, 
            'n_calls': 50, 
            'shuffle': True, 
            'early_stopping_rounds': 1,
            'fixed_parameters': training_default_params, 
            'random_state': RANDOM_STATE, 
            'use_gpu': False}
        
        optimizer = LightGBMOptimizer(**optimizer_params)
        
        params = optimizer.optimize(x_train, y_train)

        params = {**params, **training_default_params}
        
        skf = StratifiedKFold(3, shuffle=True, random_state=RANDOM_STATE)

        models = []
        
        for train_index, valid_index in skf.split(x_train, y_train):

            xx_train, yy_train = x_train.values[train_index, :], y_train.values[train_index]
            xx_valid, yy_valid = x_train.values[valid_index, :], y_train.values[valid_index]

            gbm = LGBMModel(**params)

            gbm.fit(xx_train, 
                    yy_train,
                    eval_set=[(xx_valid, yy_valid)],
                    early_stopping_rounds=1,
                    verbose=False)
            
            models.append(gbm)
        
        #################################################################################################
        # Predicting
        #################################################################################################        
        
        y_hat_train, y_hat_valid = None, None
        
        for model in models:
            
            y_ = model.predict(x_train)
            y_hat_train = y_ if y_hat_train is None else y_ + y_hat_train
            
            y_ = model.predict(x_valid)
            y_hat_valid = y_ if y_hat_valid is None else y_ + y_hat_valid
        
        y_hat_train /= len(models)
        y_hat_valid /= len(models)
        
        #################################################################################################
        # Analysing Performance
        #################################################################################################   
        
        # Computing AUC
        train_auc = roc_auc_score(y_train, y_hat_train)
        valid_auc = roc_auc_score(y_valid, y_hat_valid)
        
        # Computing logLoss
        train_loss = log_loss(y_train, y_hat_train)
        valid_loss = log_loss(y_valid, y_hat_valid)
        
        # Compute optimized threshold
        opt_threshold = optimize_threshold(y_train, y_hat_train)

        if opt_threshold is None:
            opt_threshold = np.mean(y_train)
        
        # compute confusion matrix
        tn, fp, fn, tp = confusion_matrix(y_valid, [int(y >= opt_threshold) for y in y_]).ravel()

        classification_results = classification_metrics(tn, fp, fn, tp)
        
        # add results to data frame (dict for now)
        for k in classification_results:
            if k not in result:
                result[k] = []
            result[k].append(classification_results[k])
        
        result['experiment'].append(experiment)
        result['marker'].append(t)
        result['train_auc'].append(train_auc)
        result['valid_auc'].append(valid_auc)
        result['train_loss'].append(train_loss)
        result['valid_loss'].append(valid_loss)
        result['execution_time'].append(time.time() - initial_time)
        result['threshold'].append(opt_threshold)
        
        log_message = 'Experiment #{}: '.format(experiment) + 'Train AUC: {}'.format(train_auc) + ' '
        log_message += 'Valid AUC: {}'.format(valid_auc)
        
        print(log_message)
    
    print('')
   
result = pd.DataFrame(result)

result.to_csv('output/gene/metrics_clinical_plus_genes.csv', sep=',', index=False)

result.head()

result.groupby('marker').mean()

result.groupby('marker').std()

result.groupby('marker').mean().mean()
