==> experiments/exp-mmrf/03_01_brfl_model_mult.ipynb <==
# Loading Data

from data import load_data
import numpy as np

clinical, _, genes, treatments, outcome = load_data()

print(clinical.shape)

clinical.describe().T

# Gene Expressions

genes.iloc[:8, :8]

# Treatments

treatments.head()

# THERAPY SENSITIVITY MODELLING

## Cross-validation Experiment

from pipeline import MuLT

from sklearn.model_selection import StratifiedKFold
from evaluation import optimize_threshold, classification_metrics
from sklearn.metrics import roc_auc_score, log_loss, confusion_matrix


from constants import N_FOLDS, RANDOM_STATE
from util import join_values

import lightgbm as lgb
import pickle as pkl
import pandas as pd
import numpy as np
import time
import os

# creating analyser object to compute and group 
# classification matrics grouped by training and validation
# dataset and by experiment id
# analyser = Analyser()

# Creating 10-fold CV splits stratified by treatments and outcome
kfold = StratifiedKFold(N_FOLDS, shuffle=True, random_state=RANDOM_STATE)
split = kfold.split(np.zeros(outcome.shape[0]), join_values([treatments, outcome]))

#
#
result = {c: [] for c in ['experiment', 'train_auc', 'valid_auc', 
                          'train_loss', 'valid_loss', 'execution_time', 'threshold']}

for experiment, (train_index, valid_index) in enumerate(split):
    
    initial_time = time.time()
    
    print('{}\n\n'.format(experiment))
    
    #######################################################################################################
    # Split train & valid
    #######################################################################################################
    
    response_train = outcome.iloc[train_index, 0]
    response_valid = outcome.iloc[valid_index, 0]

    clinical_train = clinical.iloc[train_index, :]
    clinical_valid = clinical.iloc[valid_index, :]

    treatments_train = treatments.iloc[train_index, :]
    treatments_valid = treatments.iloc[valid_index, :]
    
    genes_train = genes.iloc[train_index, :]
    genes_valid = genes.iloc[valid_index, :]
    
    #######################################################################################################
    # Train & Test distances
    #######################################################################################################

    if False:
        
        dists = []

        for row_train in clinical.join(genes, how='inner').values:
            for row_valid in clinical.join(genes, how='inner').values:
                dists.append(np.linalg.norm(row_train-row_valid))

        train_test_distance_avg = np.mean(dists)
        train_test_distance_std = np.std(dists)
        train_test_distance_min = np.min(dists)
        train_test_distance_max = np.max(dists)

    #######################################################################################################
    # MuLT fitting
    #######################################################################################################
    
    mult = MuLT(experiment_number=experiment, 
                number_of_experiments=N_FOLDS, 
                output_path='./output/mult/', 
                random_state=RANDOM_STATE)

    mult.fit(clinical_train, genes_train, treatments_train, response_train, 

        lgb_fixed_parameters = {
            'metric': 'binary_logloss',
            'n_estimators': 100,
            'objective': 'binary',
            'is_unbalance': False, 
            'extra_trees': True,
            'max_depth': 4,
            'learning_rate': 0.1,
            'min_split_gain': 0.0001,
            'min_child_weight': 0.0001},

        optimization_n_call=50,
        optimization_n_folds=2,
        optimization_early_stopping_rounds=1,

        clinical_marker_selection_threshold=0.050,
        gene_selection_threshold=0.002,

        dae_decay_rate=1.0,
        dae_learning_rate=1e-1,
        dae_steps=100000,
        dae_early_stopping_rounds=1000,

        lgb_early_stopping_rounds=1,

        predictor_n_folds=3)

    with open('output/mult/trained_model_{}.pkl'.format(experiment), 'wb') as file:
        pkl.dump(mult, file)
    
    #######################################################################################################
    # MuLT inference
    #######################################################################################################

    y_hat_train = mult.predict(clinical_train, genes_train, treatments_train)
    y_hat_valid = mult.predict(clinical_valid, genes_valid, treatments_valid)

    #################################################################################################
    # Analysing Performance
    #################################################################################################   

    # Computing AUC
    train_auc = roc_auc_score(response_train, y_hat_train)
    valid_auc = roc_auc_score(response_valid, y_hat_valid)

    # Computing logLoss
    train_loss = log_loss(response_train, y_hat_train)
    valid_loss = log_loss(response_valid, y_hat_valid)

    # Compute optimized threshold
    opt_threshold = optimize_threshold(response_train, y_hat_train)

    if opt_threshold is None:
        opt_threshold = np.mean(response_train)

    # compute confusion matrix
    tn, fp, fn, tp = confusion_matrix(response_valid, [int(y >= opt_threshold) for y in y_hat_valid]).ravel()

    classification_results = classification_metrics(tn, fp, fn, tp)

    # add results to data frame (dict for now)
    for k in classification_results:
        if k not in result:
            result[k] = []
        result[k].append(classification_results[k])

    result['experiment'].append(experiment)
    result['train_auc'].append(train_auc)
    result['valid_auc'].append(valid_auc)
    result['train_loss'].append(train_loss)
    result['valid_loss'].append(valid_loss)
    result['execution_time'].append(time.time() - initial_time)
    result['threshold'].append(opt_threshold)

    print('Experiment {} with {} genes and {} clinical markers'.format(
          experiment, len(mult.selected_genes[0]), len(mult.selected_clinical[0])))

    print('Train: {}'.format(train_auc))

    print('Valid: {}'.format(valid_auc))

    print("\n========================================================================================\n")
    
    # Exporting inference
    response_train = pd.DataFrame(response_train)
    response_train['y_hat'] = y_hat_train
    response_train.to_csv('output/mult/inference/train_{}.csv'.format(experiment), index=True, sep=',')
    
    response_valid = pd.DataFrame(response_valid)
    response_valid['y_hat'] = y_hat_valid
    response_valid.to_csv('output/mult/inference/valid_{}.csv'.format(experiment), index=True, sep=',')

result = pd.DataFrame(result)

result.to_csv('output/mult/metrics.csv', sep=',', index=False)

result

result.mean()

result.std()

## Analyse per Treatments

from sklearn.metrics import confusion_matrix
from sklearn.metrics import roc_auc_score
from sklearn.metrics import accuracy_score

def classification_metrics(x):
    try:
        
        tn, fp, fn, tp = confusion_matrix(x['y_true'], x['y_hat_hard']).ravel()
        
        sensitivity, specificity = tp / (tp + fn), tn / (tn + fp)
        
        precision = tp / (tp + fp)
        
        return pd.Series({
            'AUC': roc_auc_score(x['y_true'], x['y_hat_adjusted']),
            'Accuracy': accuracy_score(x['y_true'], x['y_hat_hard']),
            'Sensitivity': sensitivity,
            'Specificity': specificity,
            'Precision': precision,
            'F1': 2 * precision * sensitivity / (precision + sensitivity)})
    
    except:
        return pd.Series({
            'AUC': 0.0,
            'Accuracy': 0.0,
            'Sensitivity': 0.0,
            'Specificity': 0.0,
            'Precision': 0.0,
            'F1': 0.0})
    
BASE_DIR = 'output/mult/inference/'
    
df = None

metrics = pd.read_csv('output/mult/metrics.csv', sep=',')

for file in os.listdir(BASE_DIR):
    if 'valid' in file:
        index = file.split('_')[1].split('.')[0]
        train = pd.read_csv(os.path.join(BASE_DIR, 'train_{}.csv'.format(index)), sep=',')
        t = metrics[metrics['experiment'] == int(index)]['threshold'].values[0]
        tmp = pd.read_csv(os.path.join(BASE_DIR, file), sep=',')
        df = tmp if df is None else pd.concat([df, tmp], axis=0)
        df['y_hat_hard'] = df['y_hat'].apply(lambda x: 1 if x >= t else 0)
        
df = df.set_index('ID', drop=True)

df = df.join(treatments[treatments==1].stack().reset_index().drop(0,1).set_index('ID').rename(
    columns={'level_1': 'treatments'}))

df = df.rename(columns={'response_best_response_first_line': 'y_true'})

df.groupby('treatments').apply(classification_metrics)
