==> experiments/exp-mmrf/03_02_smla_vs_mult.ipynb <==
smla = read.csv('output/smla/metrics.csv', sep=',')

smla['X'] <- NULL

head(smla)

nmla = read.csv('output/nmla/metrics.csv', sep=',')

head(nmla)

t.test(nmla$valid_auc, smla$valid_auc, alternative='greater')

?t.test
